fixed version

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: jsaariko <jsaariko@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/10/29 14:22:04 by jsaariko       #+#    #+#                */
-/*   Updated: 2019/10/29 14:22:28 by jsaariko      ########   odam.nl         */
+/*   Updated: 2019/11/07 17:51:06 by jsaariko      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int l;
+	size_t l;
 
 	l = 0;
 	while (s[l] != '\0')
